Cleanup Change - Move scripts and css to dedicated file Change - Add directory check to file delete methods Change - Links based on account

+++ zeaploy/Services/Services/FileService.cs
@@ -5,19 +5,28 @@
         public void DeleteProfilePicture(string name)
         {
             string relativePath = $@"wwwroot\user-data\profile-pictures\{name}\";
-            Directory.Delete(relativePath, true);
+            if (Directory.Exists(relativePath))
+            {
+                Directory.Delete(relativePath, true);
+            }
         }
 
         public void DeleteCompanyLogo(string name)
         {
             string relativePath = $@"wwwroot\company-logos\{name}\";
-            Directory.Delete(relativePath, true);
+            if (Directory.Exists(relativePath))
+            {
+                Directory.Delete(relativePath, true);
+            }
         }
 
         public void DeleteApplicationFiles(string name, string companyName)
         {
             string relativePath = $@"wwwroot\user-data\cv-letters\{name}\{companyName}";
-            Directory.Delete(relativePath, true);
+            if (Directory.Exists(relativePath))
+            {
+                Directory.Delete(relativePath, true);
+            }
         }
 
         public async Task UploadApplicationFileAsync(IFormFile applicationFile, string name, string companyName)
